+++ packages/amplify-category-geo/src/service-walkthroughs/resourceWalkthrough.ts
@@ -1,7 +1,7 @@
-import { AccessType, DataProvider, ResourceParameters } from "../service-utils/resourceParams";
-import { apiDocs, ServiceName } from "../service-utils/constants";
+import { AccessType, DataProvider, ResourceParameters } from '../service-utils/resourceParams';
+import { apiDocs, ServiceName } from '../service-utils/constants';
 import { prompter, printer, byValue, byValues } from 'amplify-prompts';
-import { $TSContext } from "amplify-cli-core";
+import { $TSContext } from 'amplify-cli-core';
 
 export async function resourceAccessWalkthrough<T extends ResourceParameters & { groupPermissions: string[] }>(
     context: $TSContext,