+++ packages/amplify-category-notifications/src/commands/notifications/console.ts
@@ -8,5 +8,5 @@ export const name = 'console';
  * @param context amplify cli context
  */
 export const run = async (context: $TSContext): Promise<void> => {
-  pinpointHelper.console(context);
+  await pinpointHelper.console(context);
 };