+++ packages/amplify-migration-tests/src/__tests__/migration_tests/notifications-migration/notifications-migration.test.ts
@@ -6,7 +6,7 @@ import {
   deleteProject,
   deleteProjectDir,
 } from '@aws-amplify/amplify-e2e-core';
-import { initJSProjectWithProfile, versionCheck } from '../../../migration-helpers';
+import { initJSProjectWithProfileV4_52_0, versionCheck } from '../../../migration-helpers';
 import { getShortId } from '../../../migration-helpers/utils';
 
 describe('amplify add notifications', () => {
@@ -31,7 +31,7 @@ describe('amplify add notifications', () => {
   it('should add in app notifications if analytics added with an older version', async () => {
     expect(migrateFromVersion.v).not.toEqual(migrateToVersion.v);
 
-    await initJSProjectWithProfile(projectRoot, {}, false);
+    await initJSProjectWithProfileV4_52_0(projectRoot, {}, false);
     await addPinpointAnalytics(projectRoot, false);
 
     const settings = { resourceName: `notification${getShortId()}` };
@@ -43,7 +43,7 @@ describe('amplify add notifications', () => {
   it('should add in app notifications if analytics added and pushed with an older version', async () => {
     expect(migrateFromVersion.v).not.toEqual(migrateToVersion.v);
 
-    await initJSProjectWithProfile(projectRoot, {}, false);
+    await initJSProjectWithProfileV4_52_0(projectRoot, {}, false);
     await addPinpointAnalytics(projectRoot, false);
     await amplifyPushAuth(projectRoot, false);